+++ app/src/main/java/com/kitzapp/telegram_stats/core/appManagers/animation/AnimationManagerInterface.java
@@ -4,9 +4,9 @@ interface AnimationManagerInterface {
 
     void setNewAlpha(String keyAlphaMap, int newAlpha);
 
-    void setNewScaleY(int newScaleY);
+    void setNewScaleY(float newScaleY);
 
-    void setNewAlphaAndScaleY(String keyAlphaMap, int newAlpha, int newScaleY);
+    void setNewAlphaAndScaleY(String keyAlphaMap, int newAlpha, float newScaleY);
 
     interface ListenerForValues {
         void animNeedInvalidate(byte animType);
